stock-alarm-bot/requirements.txt
# --- 핵심 데이터 수집 및 분석 ---
pykrx==1.0.48
yfinance==0.2.52
pandas==2.2.3
OpenDartReader==1.0.23

# --- 시스템 및 유틸리티 ---
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2